requirements.txt
Flask>=2.0
moviepy
pydub
tqdm
torch>=2.0
transformers>=4.41
accelerate
sentencepiece
whisperx
pyannote.audio
numpy
python-dotenv
